Add RSS feed for PostgreSQL posts
Note that the posts are categorized as postgresql-2 internally to fit
with the old infrastructure of Wordpress that uselessly used with category
name instead of a plain "postgresql". To the users it is shown that the
posts are categorized as "postgresql", change made kind of transparent
thanks to this feed file.