Quick-and-dirty fix to bypass pagination for posts
Pagination has been introduced but it actually removed the content
for all the posts. So use a dirty fix based on the assumption that
all the posts include _posts in their path to bypass pagination in
this case and print only pagination for the front page.

This should be removed, but will make it for now as main focus is
recovery of the data.