Add new CSS layout for this blog
This style is based on a two-column layer, with left column covering 75%
of screen and right column covering 25% of screen whatever the size of the
screen. Right-side is used for reference, external and utilities while
left-side is dedicated to content, like pages and posts.

Header part is kept sober and can contain many items referencing to pages
mainly.

README is updated as well to refer to the licenses of each part of this
code, as content, site code and materials are under different licenses.