Fix paging at the bottom of home page
The base URL used for a page invocated using the list of links at the
bottom was using the current page, making two consecutive clicks on it
point to a page like 'http://site.com/pageXX/pageYY' but it does not
actually exist. The fix is rather simple, instead of using site.baseurl
that is the current page, hardcode base path to '/' in all cases. At the
same time remove the harcoded mentions to index.html, this is inelegant.

I actuallt noticed that behavior in the past, but didn't bother looking
at it further even if it has proved to be finally trivial to fix. Special
thanks to Matthew Schinckel for complaining about this broken behavior.