Remove dependency to redcarpet
It happens that kramdown is able to handle table format properly, so
just rely on that. While looking at it, add a dependency with
jekyll-paginate by listing it explicitely as a gem this blog is using
and ensure that nothing will be broken later on.

This was preventing the blog from rebuilding and refreshing itself.