Implement blog post tags without ruby dependency
Up to now, I had on this blog a plugin in Ruby which was used to
generate sets of tags for a blog with links to them, however this cannot
work on github as jekyll server is invoked with the --safe mode, in
charge of blocking custom plugins.

In this commit, the ruby plugin is replaced by a more complete in-core
solution about to generate tags without any plugins.  This costs in
creating an additional markdown file in tag/ to be able to create a tag
file for it, but at least that's a native solution, and I was looking
for one for a couple of years.

This implements the facility able to achieve tag support, are still
missing a lot of files in charge of doing the per-tag support.