Do a complete rework of the manuals
This commit is a long-overdue rework of the section dedicated to manuals
for the blog.  This mainly removes a lot of outdated information or
information that I have found no use for in the last couple of years.
Some links are renamed to have a better style.  A section dedicated to
Linux is created.

This is mostly some personal notes, so please don't complain back to me
if you have issues with this work.